[WEB-1855] - https://github.com/thegoodparty/gp-webapp/pull/156#discussion_r1572680294
